amspApp/requirements.txt
alabaster==0.7.6
amqp==2.4.0
anyjson==0.3.3
appdirs==1.4.3
arrow==0.7.0
asq==1.0
attrs==18.1.0
Babel==2.0
backports-abc==0.4
backports.functools-lru-cache==1.2.1
beautifulsoup4==4.4.0
billiard==3.5.0.5
cached-property==1.4.3
celery==4.2.1
certifi==2015.9.6.2
chardet==3.0.4
coverage==4.0.1
dateutils==0.6.6
defusedxml==0.5.0
Django==1.8.3
django-angular==0.7.15
django-appconf==1.0.2
django-bootstrap3==6.2.2
django-bootstrap3-form==0.3.0
django-browserid==1.0.0
django-classy-tags==0.6.2
django-compressor==2.0
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-el-pagination==2.1.1
django-filter==1.0.4
django-filters==0.2.1
django-http-proxy==0.4.3
django-modelcluster==1.1
django-mongodb-engine==0.6.0
django-pony-forms==0.3.9
django-related==0.0.6
django-rest-framework-mongoengine==2.0.2
django-sekizai==0.8.2
django-simple-captcha==0.4.6
django-taggit==0.18.1
django-treebeard==4.0
django-uuidfield==0.5.0
django-websocket-redis==0.4.5
djangorestframework==3.1.3
djangotoolbox==1.8.0
docutils==0.12
drf-nested-routers==0.9.0
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
fcm-django==0.2.19
futures==3.0.5
geopy==1.11.0
gevent==1.1.1
goslate==1.5.1
greenlet==0.4.9
html5lib==0.9999999
isodate==0.6.0
Jinja2==2.8
kombu==4.2.2.post1
linecache2==1.0.0
lxml==3.6.0
MarkupSafe==0.23
mockredispy==2.9.3
mongoengine==0.9.0
Monotime==1.0
mysqlclient==1.3.6
nose==1.3.7
numpy==1.12.1
oauthlib==1.0.1
olefile==0.44
pdfminer.six==20170720
pdfminer2==20151206
pika==0.10.0
Pillow==4.1.1
pluggy==0.3.0
py==1.4.30
pycares==1.0.0
pycryptodome==3.4.6
pyelasticsearch==1.4
pyfcm==1.4.5
pyfiglet==0.7.4
Pygments==2.0.2
pymongo==2.8
pymssql==2.1.2
PyMySQL==0.7.1
PyPDF2==1.26.0
PySocks==1.5.6
python-bidi==0.4.0
python-dateutil==2.6.1
python-docx==0.8.5
python-memcached==1.57
python3-pika==0.9.14
pytz==2018.9
rainbowstream==1.3.3
rcssmin==1.0.6
redis==3.1.0
requests==2.7.0
requests-oauthlib==0.5.0
requests-toolbelt==0.8.0
rjsmin==1.0.12
simplejson==3.8.0
six==1.11.0
snowballstemmer==1.2.0
sorl-thumbnail==12.3
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
SpiffWorkflow==0.4
tapioca-disqus==0.1.2
tapioca-wrapper==0.5.3
tornado==4.3
tornado-celery==0.3.5
tox==2.1.1
traceback2==1.4.0
translate==2.0.0
tweepy==3.3.0
Twisted==15.5.0
twitter==1.17.1
Unidecode==0.4.19
unittest2==1.1.0
urllib3==1.22
uWSGI==2.0.15
vine==1.2.0
virtualenv==13.1.0
wagtail==1.4.3
Wand==0.4.4
Willow==0.3
wsaccel==0.6.2
XlsxWriter==0.8.4
zeep==3.0.0
zope.interface==4.1.3
